requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.0
attrs==23.1.0
billiard==4.2.0
blinker==1.7.0
build==1.0.3
CacheControl==0.13.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.0
configobj==5.0.8
contourpy==1.2.0
crashtest==0.4.1
cryptography==41.0.7
cycler==0.12.1
databricks-cli==0.18.0
debugpy==1.8.0
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docker==6.1.3
dpath==2.1.6
dulwich==0.21.7
dvc==3.36.1
dvc-data==3.3.0
dvc-http==2.32.0
dvc-objects==3.0.0
dvc-render==1.0.0
dvc-studio-client==0.18.0
dvc-task==0.3.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.106.0
fastjsonschema==2.19.0
filelock==3.13.1
flake8==6.1.0
Flask==3.0.0
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
funcy==2.0
gitdb==4.0.11
GitPython==3.1.40
grandalf==0.8
greenlet==3.0.3
gto==1.6.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
hydra-core==1.3.2
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.28.0
ipython==8.19.0
isort==5.13.2
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.6.0
keyring==24.3.0
kiwisolver==1.4.5
kombu==5.3.4
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mlflow==2.9.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.2.1
nodeenv==1.8.0
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
orjson==3.9.10
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathspec==0.12.1
pexpect==4.9.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.3.0
poetry==1.7.1
poetry-conda==0.1.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==1.4.2
pyflakes==3.1.0
pygit2==1.13.3
Pygments==2.17.2
pygtrie==2.5.0
PyJWT==2.8.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyright==1.1.343
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
querystring-parser==1.2.4
rapidfuzz==3.6.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-learn==1.3.2
scipy==1.11.4
scmrepo==2.0.2
SecretStorage==3.3.3
semver==3.0.2
shellingham==1.5.4
shortuuid==1.0.11
shtab==1.6.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
sqltrie==0.11.0
stack-data==0.6.3
starlette==0.27.0
tabulate==0.9.0
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.3
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
trove-classifiers==2023.11.29
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
vine==5.1.0
virtualenv==20.25.0
voluptuous==0.14.1
wcwidth==0.2.12
websocket-client==1.7.0
Werkzeug==3.0.1
yapf==0.40.2
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0
